+++ Reflection/src/utopia/reflection/component/swing/button/FramedImageButton.scala
@@ -5,6 +5,7 @@ import utopia.reflection.component.context.ButtonContextLike
 import utopia.reflection.component.drawing.mutable.CustomDrawableWrapper
 import utopia.reflection.component.swing.label.ImageLabel
 import utopia.reflection.component.swing.template.{StackableAwtComponentWrapperWrapper, SwingComponentRelated}
+import utopia.reflection.event.ButtonState
 import utopia.reflection.image.SingleColorIcon
 import utopia.reflection.shape.LengthExtensions._
 import utopia.reflection.shape.stack.StackInsets