+++ Reflection/src/utopia/reflection/component/swing/button/ImageButton.scala
@@ -3,6 +3,7 @@ package utopia.reflection.component.swing.button
 import utopia.reflection.component.context.BaseContextLike
 import utopia.reflection.component.swing.label.ImageLabel
 import utopia.reflection.component.swing.template.{StackableAwtComponentWrapperWrapper, SwingComponentRelated}
+import utopia.reflection.event.ButtonState
 
 object ImageButton
 {