+++ Reflection/src/utopia/reflection/component/swing/button/ImageCheckBox.scala
@@ -5,6 +5,7 @@ import utopia.reflection.component.drawing.mutable.CustomDrawableWrapper
 import utopia.reflection.component.template.input.InteractionWithPointer
 import utopia.reflection.component.swing.label.ImageLabel
 import utopia.reflection.component.swing.template.{StackableAwtComponentWrapperWrapper, SwingComponentRelated}
+import utopia.reflection.event.ButtonState
 
 /**
   * Used for switching a value on or off